+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2019-07-17
+### Fixed
+- [Changed the phone number format for Sweden](https://github.com/alexplumb/material-ui-phone-number/pull/19)
+- Update lodash for security vulnerability
+- [Removed reliance on theme providers](https://github.com/alexplumb/material-ui-phone-number/issues/18)
+
 ## [2.0.0-beta0] - 2019-05-02
 ### Added
 - Added support for a new `native` property for people who want to use a native select for the button dropdown